Crate aws_structured_encryption

This is an awesome crate

Structs

AuthItem

One piece of data to be decrypted

CryptoItem

One piece of data to be encrypted

DecryptInput

Input for decrypt

DecryptOutput

Output of decrypt

EncryptInput

Input for encrypt

EncryptOutput

Output of encrypt

Error

Base error type for aws_structured_encryption

ParsedHeader

Parsed Header

ResolveActionsInput

Examine the header to determine which attributes were encrypted

ResolveActionsOutput

The result of resolve_actions

StructureSegment

A part of a Path that is a member of a structure

StructuredDataTerminal

The actual data to be encrypted or decrypted

Enums

AuthenticateAction

To sign or not to sign

CryptoAction

The cryptographic action to take on a StructuredDataTerminal

ErrorKind

Individual error types for aws_esdk

PathSegment

one part of a path

Constants

TYPE_ID_LEN

Size of TerminalTypeId

Functions

decrypt

Decrypt the structure

encrypt

Encrypt the structure

resolve_actions

For each attribute, is it encrypted?

Type Aliases

TerminalTypeId

A 2 byte value that is used to identify how callers interpret a terminal value